src/requirements.txt
Django==5.2.10
boto3==1.40.68
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
